+++ src/accordion/accordion.ts
@@ -127,14 +127,12 @@ export interface NgbPanelChangeEvent {
     <ng-template ngFor let-panel [ngForOf]="panels">
       <div class="card">
         <div role="tab" id="{{panel.id}}-header" [class]="'card-header ' + (panel.type ? 'bg-'+panel.type: type ? 'bg-'+type : '')">
-          <h5 class="mb-0">
-            <button type="button" class="btn btn-link"
-              (click)="toggle(panel.id)" [disabled]="panel.disabled" [class.collapsed]="!panel.isOpen"
-              [attr.aria-expanded]="panel.isOpen" [attr.aria-controls]="panel.id">
-              {{panel.title}}<ng-template [ngTemplateOutlet]="panel.titleTpl?.templateRef"></ng-template>
-            </button>
-            <ng-template [ngTemplateOutlet]='panel.titleControlTpl?.templateRef'></ng-template>
-          </h5>
+          <button type="button" class="btn btn-link"
+            (click)="toggle(panel.id)" [disabled]="panel.disabled" [class.collapsed]="!panel.isOpen"
+            [attr.aria-expanded]="panel.isOpen" [attr.aria-controls]="panel.id">
+            {{panel.title}}<ng-template [ngTemplateOutlet]="panel.titleTpl?.templateRef"></ng-template>
+          </button>
+          <ng-template [ngTemplateOutlet]='panel.titleControlTpl?.templateRef'></ng-template>
         </div>
         <div id="{{panel.id}}" role="tabpanel" [attr.aria-labelledby]="panel.id + '-header'"
              class="collapse" [class.show]="panel.isOpen" *ngIf="!destroyOnHide || panel.isOpen">