+++ src/datepicker/datepicker-keymap-service.ts
@@ -1,7 +1,6 @@
 import {Injectable} from '@angular/core';
 import {NgbDatepickerService} from './datepicker-service';
 import {NgbCalendar} from './ngb-calendar';
-import {toString} from '../util/util';
 import {Key} from '../util/key';
 import {NgbDate} from './ngb-date';
 
@@ -23,43 +22,41 @@ export class NgbDatepickerKeyMapService {
 
   processKey(event: KeyboardEvent) {
     // tslint:disable-next-line:deprecation
-    const {which} = event;
-    if (Key[toString(which)]) {
-      switch (which) {
-        case Key.PageUp:
-          this._service.focusMove(event.shiftKey ? 'y' : 'm', -1);
-          break;
-        case Key.PageDown:
-          this._service.focusMove(event.shiftKey ? 'y' : 'm', 1);
-          break;
-        case Key.End:
-          this._service.focus(event.shiftKey ? this._maxDate : this._lastViewDate);
-          break;
-        case Key.Home:
-          this._service.focus(event.shiftKey ? this._minDate : this._firstViewDate);
-          break;
-        case Key.ArrowLeft:
-          this._service.focusMove('d', -1);
-          break;
-        case Key.ArrowUp:
-          this._service.focusMove('d', -this._calendar.getDaysPerWeek());
-          break;
-        case Key.ArrowRight:
-          this._service.focusMove('d', 1);
-          break;
-        case Key.ArrowDown:
-          this._service.focusMove('d', this._calendar.getDaysPerWeek());
-          break;
-        case Key.Enter:
-        case Key.Space:
-          this._service.focusSelect();
-          break;
-        default:
-          return;
-      }
-
-      event.preventDefault();
-      event.stopPropagation();
+    switch (event.which) {
+      case Key.PageUp:
+        this._service.focusMove(event.shiftKey ? 'y' : 'm', -1);
+        break;
+      case Key.PageDown:
+        this._service.focusMove(event.shiftKey ? 'y' : 'm', 1);
+        break;
+      case Key.End:
+        this._service.focus(event.shiftKey ? this._maxDate : this._lastViewDate);
+        break;
+      case Key.Home:
+        this._service.focus(event.shiftKey ? this._minDate : this._firstViewDate);
+        break;
+      case Key.ArrowLeft:
+        this._service.focusMove('d', -1);
+        break;
+      case Key.ArrowUp:
+        this._service.focusMove('d', -this._calendar.getDaysPerWeek());
+        break;
+      case Key.ArrowRight:
+        this._service.focusMove('d', 1);
+        break;
+      case Key.ArrowDown:
+        this._service.focusMove('d', this._calendar.getDaysPerWeek());
+        break;
+      case Key.Enter:
+      case Key.Space:
+        this._service.focusSelect();
+        break;
+      default:
+        return;
     }
+
+    // note 'return' in default case
+    event.preventDefault();
+    event.stopPropagation();
   }
 }